refactor(i18n): isolate backend locale synchronization
- Move native locale synchronization into a dedicated module
- Route synchronization failures through the shared error handler
- Keep the frontend i18n module free of circular dependencies
