+++ src/i18n/index.ts
@@ -1,4 +1,3 @@
-import { invoke } from "@tauri-apps/api/core";
 import i18n from "i18next";
 import { initReactI18next } from "react-i18next";
 import { enUS } from "./locales/en-us";
@@ -31,29 +30,4 @@ i18n.on("languageChanged", (language) => {
 	localStorage.setItem(LANGUAGE_STORAGE_KEY, language);
 });
 
-/**
- * Synchronizes one i18next language change with the native error translation boundary.
- *
- * @example
- * backendLanguageChanged("zh-CN");
- */
-const backendLanguageChanged = (language: string) => {
-	invoke("set_backend_locale", { locale: language }).catch((error) => {
-		console.error("Failed to synchronize the backend locale", error);
-	});
-};
-
-/**
- * Initializes native IPC error messages with the active i18next language.
- *
- * @example
- * initializeBackendI18n();
- */
-const initializeBackendI18n = () => {
-	backendLanguageChanged(i18n.resolvedLanguage ?? i18n.language);
-};
-
-i18n.on("languageChanged", backendLanguageChanged);
-
-export { initializeBackendI18n };
 export default i18n;
