+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import { initializeBackendI18n } from "./i18n";
+import { initializeBackendI18n } from "./i18n/backend";
 import "./styles.css";
 
 initializeBackendI18n();
